requirements.txt
aiofiles==23.2.1
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.3
blinker==1.7.0
boto3==1.34.89
botocore==1.34.89
certifi==2024.2.2
click==8.1.7
databases==0.9.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.110.2
fastapi-storages==0.3.0
Flask==3.0.3
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.3
MarkupSafe==2.1.5
orjson==3.10.2
passlib==1.7.4
peewee==3.17.3
Pillow==9.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqladmin==0.16.1
SQLAlchemy==2.0.29
sqlmodel==0.0.16
starlette==0.37.2
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
wtf==0.1
WTForms==3.1.2
